LayoutTests/accessibility/gtk/text-for-range-entry-and-password-expected.txt


This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entry.stringForRange(0, -1) is 'foo bar baz'
PASS password.stringForRange(0, -1) is '•••••••••••'
PASS successfullyParsed is true

TEST COMPLETE
